requirements.txt
CairoSVG==2.7.1
Jinja2==3.1.2
PyYAML==6.0.1
yq==3.4.3